2.4 Debug. Finish semantic analysis. Almost pass all the semantic test cases.

+++ src/MxCompiler/AST/IntLiteralNode.java
@@ -3,14 +3,14 @@ package MxCompiler.AST;
 import MxCompiler.Utilities.Location;
 
 public class IntLiteralNode extends ConstExprNode {
-    private int value;
+    private long value;
 
-    public IntLiteralNode(Location location, String text, int value) {
+    public IntLiteralNode(Location location, String text, long value) {
         super(location, text);
         this.value = value;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
